chore(sync): 同步上游客户端修复批次
从 DDNet 上游 206 项语义审查收敛客户端批次（2026-08-21）。

## fix
- 修复 OnNewSnapshot 在 dummy 切换后未刷新相机（新增 DummySwapped 参数）
- 修复 demo 播放 delta 快照前无 full snapshot 时的崩溃
- 修复短 demo 下跳转时长索引越界（Shift+方向键）
- 修复时间渲染 absolute(INT_MIN) 未定义行为（hud/ui）
- 修复 tiles 图层数据访问未校验（空指针/尺寸越界）
- 修复截图创建失败无错误日志
- 修复 video.cpp 变量名 typo（MakeWriteable）

## refactor
- IClient::ServerInfo() 替换 15 处 GetServerInfo 拷贝调用（chat/statboard/menus/tclient 等）
- Console()->Print 迁移 log_*（broadcast/motd/chat/menus/gameclient），新增 base/log_color.h 的 color_cast<LOG_COLOR>
- EscapeUrl 参数 int 改 size_t，调用点用 str_startswith 与模板版
- 删除未使用的 CMenus::m_SettingPlayerPage

验证：改动文件 compile_commands 编译通过；quick gate 11/0/0。

+++ src/engine/shared/http.cpp
@@ -140,7 +140,7 @@ static std::string HttpRequestHostKey(const char *pUrl)
 	return Host;
 }
 
-void EscapeUrl(char *pBuf, int Size, const char *pStr)
+void EscapeUrl(char *pBuf, size_t Size, const char *pStr)
 {
 	char *pEsc = curl_easy_escape(nullptr, pStr, 0);
 	str_copy(pBuf, pEsc, Size);
