+++ src/engine/shared/http.h
@@ -323,7 +323,7 @@ inline std::unique_ptr<CHttpRequest> HttpPostJson(const char *pUrl, const char *
 	return pResult;
 }
 
-void EscapeUrl(char *pBuf, int Size, const char *pStr);
+void EscapeUrl(char *pBuf, size_t Size, const char *pStr);
 
 template<int N>
 void EscapeUrl(char (&aBuf)[N], const char *pStr)
